исправила регистрацию, добавление юзера в бд
