+++ application/admin.py
@@ -1,0 +1,7 @@
+from django.contrib import admin
+
+from django.contrib import admin
+
+from .models import User
+
+admin.site.register(User)
