+++ application/serializer.py
@@ -1,4 +1,4 @@
-from application.models import User
+from .models import User
 from rest_framework import serializers
