+++ application/templates/application/signup.html
@@ -1,0 +1,18 @@
+{% extends 'application/layout/base.html' %}
+
+
+{% block content %}
+    <form method="post">
+        <div>
+            <input required name="email" type="text">
+        </div>
+        <div>
+            <input required name="name" type="text">
+        </div>
+        <div>
+            <input required name="pass" type="password">
+        </div>
+        <input type="submit">
+        {% csrf_token %}
+    </form>
+{% endblock %}
