+++ application/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+
+from application import views
+
+urlpatterns = [
+    path('signup/', views.show_signup)
+]
